requirements.txt
langgraph>=0.0.40
pydantic>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
uvicorn>=0.23.0
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0
python-multipart>=0.0.6
click>=8.1.0
langgraph>=0.2.0
matplotlib